extract training setup bolier plate code
