+++ lib/api/unittest/CDataFrameAnalyzerTrainingTest.cc
@@ -20,10 +20,10 @@
 #include <api/CSingleStreamDataAdder.h>
 #include <api/ElasticsearchStateIndex.h>
 
-#include <test/CDataFrameAnalysisSpecificationFactory.h>
-#include <test/CRandomNumbers.h>
-
 #include <test/BoostTestCloseAbsolute.h>
+#include <test/CDataFrameAnalysisSpecificationFactory.h>
+#include <test/CDataFrameAnalyzerTrainingFactory.h>
+#include <test/CRandomNumbers.h>
 
 #include <boost/test/unit_test.hpp>
 #include <boost/unordered_map.hpp>
@@ -101,187 +101,6 @@ auto restoreTree(std::string persistedState, TDataFrameUPtr& frame, std::size_t 
         *frame, dependentVariable);
 }
 
-TDataFrameUPtr setupLinearRegressionData(const TStrVec& fieldNames,
-                                         TStrVec& fieldValues,
-                                         api::CDataFrameAnalyzer& analyzer,
-                                         const TDoubleVec& weights,
-                                         const TDoubleVec& regressors,
-                                         TStrVec& targets) {
-
-    auto target = [&weights](const TDoubleVec& regressors_) {
-        double result{0.0};
-        for (std::size_t i = 0; i < weights.size(); ++i) {
-            result += weights[i] * regressors_[i];
-        }
-        return core::CStringUtils::typeToStringPrecise(result, core::CIEEE754::E_DoublePrecision);
-    };
-
-    auto frame = core::makeMainStorageDataFrame(weights.size() + 1).first;
-
-    for (std::size_t i = 0; i < regressors.size(); i += weights.size()) {
-        TDoubleVec row(weights.size());
-        for (std::size_t j = 0; j < weights.size(); ++j) {
-            row[j] = regressors[i + j];
-        }
-
-        for (std::size_t j = 0; j < row.size(); ++j) {
-            fieldValues[j] = core::CStringUtils::typeToStringPrecise(
-                row[j], core::CIEEE754::E_DoublePrecision);
-        }
-        fieldValues[weights.size()] = target(row);
-        targets.push_back(fieldValues[weights.size()]);
-
-        analyzer.handleRecord(fieldNames, fieldValues);
-        frame->parseAndWriteRow(
-            core::CVectorRange<const TStrVec>(fieldValues, 0, weights.size() + 1));
-    }
-
-    frame->finishWritingRows();
-
-    return frame;
-}
-
-TDataFrameUPtr setupBinaryClassificationData(const TStrVec& fieldNames,
-                                             TStrVec& fieldValues,
-                                             api::CDataFrameAnalyzer& analyzer,
-                                             const TDoubleVec& weights,
-                                             const TDoubleVec& regressors,
-                                             TStrVec& targets) {
-    TStrVec classes{"foo", "bar"};
-    auto target = [&weights, &classes](const TDoubleVec& regressors_) {
-        double result{0.0};
-        for (std::size_t i = 0; i < weights.size(); ++i) {
-            result += weights[i] * regressors_[i];
-        }
-        return classes[result < 0.0 ? 0 : 1];
-    };
-
-    auto frame = core::makeMainStorageDataFrame(weights.size() + 1).first;
-    TBoolVec categoricalFields(weights.size(), false);
-    categoricalFields.push_back(true);
-    frame->categoricalColumns(std::move(categoricalFields));
-
-    for (std::size_t i = 0; i < regressors.size(); i += weights.size()) {
-        TDoubleVec row(weights.size());
-        for (std::size_t j = 0; j < weights.size(); ++j) {
-            row[j] = regressors[i + j];
-        }
-
-        for (std::size_t j = 0; j < row.size() - 1; ++j) {
-            fieldValues[j] = core::CStringUtils::typeToStringPrecise(
-                row[j], core::CIEEE754::E_DoublePrecision);
-        }
-        fieldValues[weights.size()] = target(row);
-        targets.push_back(fieldValues[weights.size()]);
-
-        analyzer.handleRecord(fieldNames, fieldValues);
-        frame->parseAndWriteRow(
-            core::CVectorRange<const TStrVec>(fieldValues, 0, weights.size() + 1));
-    }
-
-    frame->finishWritingRows();
-
-    return frame;
-}
-
-enum EPredictionType { E_Regression, E_BinaryClassification };
-
-void appendPrediction(core::CDataFrame&, std::size_t, double prediction, double, TDoubleVec& predictions) {
-    predictions.push_back(prediction);
-}
-
-void appendPrediction(core::CDataFrame& frame,
-                      std::size_t columnHoldingPrediction,
-                      double logOddsClass1,
-                      double threshold,
-                      TStrVec& predictions) {
-    predictions.push_back(
-        maths::CTools::logisticFunction(logOddsClass1) < threshold
-            ? frame.categoricalColumnValues()[columnHoldingPrediction][0]
-            : frame.categoricalColumnValues()[columnHoldingPrediction][1]);
-}
-
-template<typename T>
-void addPredictionTestData(EPredictionType type,
-                           const TStrVec& fieldNames,
-                           TStrVec fieldValues,
-                           api::CDataFrameAnalyzer& analyzer,
-                           std::vector<T>& expectedPredictions,
-                           std::size_t numberExamples = 100,
-                           double alpha = -1.0,
-                           double lambda = -1.0,
-                           double gamma = -1.0,
-                           double softTreeDepthLimit = -1.0,
-                           double softTreeDepthTolerance = -1.0,
-                           double eta = 0.0,
-                           std::size_t maximumNumberTrees = 0,
-                           double featureBagFraction = 0.0) {
-
-    test::CRandomNumbers rng;
-
-    TDoubleVec weights;
-    rng.generateUniformSamples(-1.0, 1.0, fieldNames.size() - 3, weights);
-    TDoubleVec regressors;
-    rng.generateUniformSamples(-10.0, 10.0, weights.size() * numberExamples, regressors);
-
-    TStrVec targets;
-    auto frame = type == E_Regression
-                     ? setupLinearRegressionData(fieldNames, fieldValues, analyzer,
-                                                 weights, regressors, targets)
-                     : setupBinaryClassificationData(fieldNames, fieldValues, analyzer,
-                                                     weights, regressors, targets);
-
-    std::unique_ptr<maths::boosted_tree::CLoss> loss;
-    if (type == E_Regression) {
-        loss = std::make_unique<maths::boosted_tree::CMse>();
-    } else {
-        loss = std::make_unique<maths::boosted_tree::CBinomialLogistic>();
-    }
-
-    maths::CBoostedTreeFactory treeFactory{
-        maths::CBoostedTreeFactory::constructFromParameters(1, std::move(loss))};
-    if (alpha >= 0.0) {
-        treeFactory.depthPenaltyMultiplier(alpha);
-    }
-    if (lambda >= 0.0) {
-        treeFactory.leafWeightPenaltyMultiplier(lambda);
-    }
-    if (gamma >= 0.0) {
-        treeFactory.treeSizePenaltyMultiplier(gamma);
-    }
-    if (softTreeDepthLimit >= 0.0) {
-        treeFactory.softTreeDepthLimit(softTreeDepthLimit);
-    }
-    if (softTreeDepthTolerance >= 0.0) {
-        treeFactory.softTreeDepthTolerance(softTreeDepthTolerance);
-    }
-    if (eta > 0.0) {
-        treeFactory.eta(eta);
-    }
-    if (maximumNumberTrees > 0) {
-        treeFactory.maximumNumberTrees(maximumNumberTrees);
-    }
-    if (featureBagFraction > 0.0) {
-        treeFactory.featureBagFraction(featureBagFraction);
-    }
-
-    ml::api::CDataFrameTrainBoostedTreeInstrumentation instrumentation("testJob");
-    treeFactory.analysisInstrumentation(instrumentation);
-
-    auto tree = treeFactory.buildFor(*frame, weights.size());
-
-    tree->train();
-    tree->predict();
-
-    frame->readRows(1, [&](TRowItr beginRows, TRowItr endRows) {
-        for (auto row = beginRows; row != endRows; ++row) {
-            double prediction{(*row)[tree->columnHoldingPrediction()]};
-            appendPrediction(*frame, weights.size(), prediction,
-                             tree->probabilityAtWhichToAssignClassOne(), expectedPredictions);
-        }
-    });
-}
-
 template<typename F>
 void testOneRunOfBoostedTreeTrainingWithStateRecovery(F makeSpec, std::size_t iterationToRestartFrom) {
 
@@ -311,8 +130,8 @@ void testOneRunOfBoostedTreeTrainingWithStateRecovery(F makeSpec, std::size_t it
                                   fieldNames.begin());
 
     TStrVec targets;
-    auto frame = setupLinearRegressionData(fieldNames, fieldValues, analyzer,
-                                           weights, regressors, targets);
+    auto frame = test::CDataFrameAnalyzerTrainingFactory::setupLinearRegressionData(
+        fieldNames, fieldValues, analyzer, weights, regressors, targets);
     analyzer.handleRecord(fieldNames, {"", "", "", "", "", "", "$"});
 
     TStrVec persistedStates{
@@ -332,8 +151,8 @@ void testOneRunOfBoostedTreeTrainingWithStateRecovery(F makeSpec, std::size_t it
         makeSpec("target", numberExamples, persisterSupplier), outputWriterFactory};
 
     targets.clear();
-    setupLinearRegressionData(fieldNames, fieldValues, restoredAnalyzer,
-                              weights, regressors, targets);
+    test::CDataFrameAnalyzerTrainingFactory::setupLinearRegressionData(
+        fieldNames, fieldValues, restoredAnalyzer, weights, regressors, targets);
     restoredAnalyzer.handleRecord(fieldNames, {"", "", "", "", "", "", "$"});
 
     persistedStates = splitOnNull(std::stringstream{std::move(persistenceStream->str())});
@@ -386,7 +205,9 @@ BOOST_AUTO_TEST_CASE(testRunBoostedTreeRegressionTraining) {
         test::CDataFrameAnalysisSpecificationFactory::predictionSpec(
             test::CDataFrameAnalysisSpecificationFactory::regression(), "target"),
         outputWriterFactory};
-    addPredictionTestData(E_Regression, fieldNames, fieldValues, analyzer, expectedPredictions);
+    test::CDataFrameAnalyzerTrainingFactory::addPredictionTestData(
+        test::CDataFrameAnalyzerTrainingFactory::E_Regression, fieldNames,
+        fieldValues, analyzer, expectedPredictions);
 
     core::CStopWatch watch{true};
     analyzer.handleRecord(fieldNames, {"", "", "", "", "", "", "$"});
@@ -450,7 +271,9 @@ BOOST_AUTO_TEST_CASE(testRunBoostedTreeRegressionTrainingStateReport) {
     api::CDataFrameAnalyzer analyzer{
         test::CDataFrameAnalysisSpecificationFactory::predictionSpec("regression", "c5"),
         outputWriterFactory};
-    addPredictionTestData(E_Regression, fieldNames, fieldValues, analyzer, expectedPredictions);
+    test::CDataFrameAnalyzerTrainingFactory::addPredictionTestData(
+        test::CDataFrameAnalyzerTrainingFactory::E_Regression, fieldNames,
+        fieldValues, analyzer, expectedPredictions);
     analyzer.handleRecord(fieldNames, {"", "", "", "", "", "", "$"});
 
     rapidjson::Document results;
@@ -488,10 +311,10 @@ BOOST_AUTO_TEST_CASE(testRunBoostedTreeRegressionTrainingWithParams) {
 
     TStrVec fieldNames{"f1", "f2", "f3", "f4", "target", ".", "."};
     TStrVec fieldValues{"", "", "", "", "", "0", ""};
-    addPredictionTestData(E_Regression, fieldNames, fieldValues, analyzer,
-                          expectedPredictions, 100, alpha, lambda, gamma,
-                          softTreeDepthLimit, softTreeDepthTolerance, eta,
-                          maximumNumberTrees, featureBagFraction);
+    test::CDataFrameAnalyzerTrainingFactory::addPredictionTestData(
+        test::CDataFrameAnalyzerTrainingFactory::E_Regression, fieldNames, fieldValues,
+        analyzer, expectedPredictions, 100, alpha, lambda, gamma, softTreeDepthLimit,
+        softTreeDepthTolerance, eta, maximumNumberTrees, featureBagFraction);
     analyzer.handleRecord(fieldNames, {"", "", "", "", "", "", "$"});
 
     // Check best hyperparameters
@@ -675,8 +498,9 @@ BOOST_AUTO_TEST_CASE(testRunBoostedTreeClassifierTraining) {
             test::CDataFrameAnalysisSpecificationFactory::classification(),
             "target", 100, 5, 6000000, 0, 0, {"target"}),
         outputWriterFactory};
-    addPredictionTestData(E_BinaryClassification, fieldNames, fieldValues,
-                          analyzer, expectedPredictions);
+    test::CDataFrameAnalyzerTrainingFactory::addPredictionTestData(
+        test::CDataFrameAnalyzerTrainingFactory::E_BinaryClassification,
+        fieldNames, fieldValues, analyzer, expectedPredictions);
 
     core::CStopWatch watch{true};
     analyzer.handleRecord(fieldNames, {"", "", "", "", "", "", "$"});
@@ -763,8 +587,8 @@ BOOST_AUTO_TEST_CASE(testRunBoostedTreeClassifierImbalanced) {
         outputWriterFactory};
 
     TStrVec actuals;
-    setupBinaryClassificationData(fieldNames, fieldValues, analyzer, weights,
-                                  regressors, actuals);
+    test::CDataFrameAnalyzerTrainingFactory::setupBinaryClassificationData(
+        fieldNames, fieldValues, analyzer, weights, regressors, actuals);
     analyzer.handleRecord(fieldNames, {"", "", "", "", "", "$"});
 
     rapidjson::Document results;
